bot-da-lista/pyproject.toml
[project]
name = "bot-da-lista"
version = "0.1.0"
description = ""
authors = [
    {name = "Rodrigololr",email = "rodrigotenorio2005@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py (>=2.5.2,<3.0.0)"
]

[tool.poetry]
packages = [{include = "bot_da_lista", from = "src"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
